services/backend/user_management/requirements.txt
Django==5.1.1
psycopg
django-crispy-forms
crispy-bootstrap5
djangorestframework
requests
python-dotenv
Pillow
djangorestframework_simplejwt
pyotp
qrcode